+++ lang-haskell/app-view/comp-view/FlareTiming/Task/Score/Effort.hs
@@ -29,14 +29,14 @@ import WireTypes.Point
     )
 import WireTypes.ValidityWorking (ValidityWorking(..), TimeValidityWorking(..))
 import WireTypes.Comp (UtcOffset(..), Discipline(..), MinimumDistance(..))
-import WireTypes.Pilot (Pilot(..), Dnf(..), DfNoTrack(..))
+import WireTypes.Pilot (Pilot(..), Dnf(..), DfNoTrack(..), pilotIdsWidth)
 import WireTypes.Effort
     ( TaskLanding(..), Lookahead(..)
     , Chunking(..), IxChunk(..), Chunk(..)
     , ChunkDifficulty(..), SumOfDifficulty(..)
     )
 import qualified WireTypes.Pilot as Pilot (DfNoTrackPilot(..))
-import FlareTiming.Pilot (showPilot, showPilotName)
+import FlareTiming.Pilot (showPilot, hashIdHyphenPilot)
 import FlareTiming.Task.Score.Show
 
 tableScoreEffort
@@ -59,6 +59,7 @@ tableScoreEffort
     -> Dynamic t (Maybe TaskLanding)
     -> m ()
 tableScoreEffort utcOffset hgOrPg free sgs ln dnf' dfNt _vy vw _wg pt _tp sDfs sEx lg' lgN' = do
+    let w = ffor sDfs (pilotIdsWidth . fmap fst)
     let dnf = unDnf <$> dnf'
     lenDnf :: Int <- sample . current $ length <$> dnf
     lenDfs :: Int <- sample . current $ length <$> sDfs
@@ -122,7 +123,7 @@ tableScoreEffort utcOffset hgOrPg free sgs ln dnf' dfNt _vy vw _wg pt _tp sDfs s
 
             el "tr" $ do
                 elClass "th" "th-placing" $ text "Place"
-                elClass "th" "th-pilot" $ text "###-Pilot"
+                elClass "th" "th-pilot" . dynText $ ffor w hashIdHyphenPilot
 
                 elClass "th" "th-chunk" $ text "Chunk"
                 elClass "th" "th-norm th-chunk" $ text "✓"
@@ -163,6 +164,7 @@ tableScoreEffort utcOffset hgOrPg free sgs ln dnf' dfNt _vy vw _wg pt _tp sDfs s
                 simpleList
                     (sortBy cmpEffort <$> sDfs)
                     (pointRow
+                        w
                         utcOffset
                         free
                         ln
@@ -173,7 +175,7 @@ tableScoreEffort utcOffset hgOrPg free sgs ln dnf' dfNt _vy vw _wg pt _tp sDfs s
                         (Map.fromList <$> pChunksN)
                     )
 
-            dnfRows dnfPlacing dnf'
+            dnfRows w dnfPlacing dnf'
             return ()
 
         let tdFoot = elAttr "td" ("colspan" =: "21")
@@ -237,7 +239,8 @@ tableScoreEffort utcOffset hgOrPg free sgs ln dnf' dfNt _vy vw _wg pt _tp sDfs s
 
 pointRow
     :: MonadWidget t m
-    => Dynamic t UtcOffset
+    => Dynamic t Int
+    -> Dynamic t UtcOffset
     -> Dynamic t MinimumDistance
     -> Dynamic t (Maybe TaskLength)
     -> Dynamic t DfNoTrack
@@ -247,7 +250,7 @@ pointRow
     -> Dynamic t (Map.Map Pilot IxChunk)
     -> Dynamic t (Pilot, Bk.Breakdown)
     -> m ()
-pointRow _utcOffset free _ln dfNt pt sEx ixChunkMap ixChunkMapN x = do
+pointRow w _utcOffset free _ln dfNt pt sEx ixChunkMap ixChunkMapN x = do
     MinimumDistance free' <- sample . current $ free
 
     let pilot = fst <$> x
@@ -257,8 +260,8 @@ pointRow _utcOffset free _ln dfNt pt sEx ixChunkMap ixChunkMapN x = do
     let points = Bk.breakdown <$> xB
 
     (classPilot, idNamePilot) <- sample . current $
-            ffor2 pilot dfNt (\p (DfNoTrack ps) ->
-                let sp = showPilot p
+            ffor3 w pilot dfNt (\w' p (DfNoTrack ps) ->
+                let sp = showPilot w' p
                 in
                     if p `elem` (Pilot.pilot <$> ps)
                        then ("pilot-dfnt", sp <> " ☞ ")
@@ -338,10 +341,11 @@ pointRow _utcOffset free _ln dfNt pt sEx ixChunkMap ixChunkMapN x = do
 
 dnfRows
     :: MonadWidget t m
-    => TaskPlacing
+    => Dynamic t Int
+    -> TaskPlacing
     -> Dynamic t Dnf
     -> m ()
-dnfRows place ps' = do
+dnfRows w place ps' = do
     let ps = unDnf <$> ps'
     len <- sample . current $ length <$> ps
     let p1 = take 1 <$> ps
@@ -351,20 +355,21 @@ dnfRows place ps' = do
         0 -> do
             return ()
         1 -> do
-            _ <- simpleList ps (dnfRow place (Just 1))
+            _ <- simpleList ps (dnfRow w place (Just 1))
             return ()
         n -> do
-            _ <- simpleList p1 (dnfRow place (Just n))
-            _ <- simpleList pN (dnfRow place Nothing)
+            _ <- simpleList p1 (dnfRow w place (Just n))
+            _ <- simpleList pN (dnfRow w place Nothing)
             return ()
 
 dnfRow
     :: MonadWidget t m
-    => TaskPlacing
+    => Dynamic t Int
+    -> TaskPlacing
     -> Maybe Int
     -> Dynamic t Pilot
     -> m ()
-dnfRow place rows pilot = do
+dnfRow w place rows pilot = do
     let dnfMega =
             case rows of
                 Nothing -> return ()
@@ -380,6 +385,6 @@ dnfRow place rows pilot = do
 
     elClass "tr" "tr-dnf" $ do
         elClass "td" "td-placing" . text $ showRank place
-        elClass "td" "td-pilot" . dynText $ showPilotName <$> pilot
+        elClass "td" "td-pilot" . dynText $ ffor2 w pilot showPilot
         dnfMega
         return ()