+++ lang-haskell/app-view/comp-view/FlareTiming/Task/Score/Penal.hs
@@ -31,9 +31,9 @@ import WireTypes.Comp
     ( Discipline(..), EarlyStart(..), JumpTheGunLimit(..)
     , showEarlyStartEarliest, showEarlyStartPenaltyRate
     )
-import WireTypes.Pilot (Pilot(..), Dnf(..), DfNoTrack(..))
+import WireTypes.Pilot (Pilot(..), Dnf(..), DfNoTrack(..), pilotIdsWidth)
 import qualified WireTypes.Pilot as Pilot (DfNoTrackPilot(..))
-import FlareTiming.Pilot (showPilot, classOfEarlyStart)
+import FlareTiming.Pilot (showPilot, hashIdHyphenPilot, classOfEarlyStart)
 import FlareTiming.Task.Score.Show
 
 tableScorePenal
@@ -53,6 +53,7 @@ tableScorePenal
     -> Dynamic t [(Pilot, Norm.NormBreakdown)]
     -> m ()
 tableScorePenal hgOrPg early sgs _ln dnf' dfNt _vy vw _wg pt tp sDfs sEx = do
+    let w = ffor sDfs (pilotIdsWidth . fmap fst)
     let dnf = unDnf <$> dnf'
     lenDnf :: Int <- sample . current $ length <$> dnf
     lenDfs :: Int <- sample . current $ length <$> sDfs
@@ -119,7 +120,7 @@ tableScorePenal hgOrPg early sgs _ln dnf' dfNt _vy vw _wg pt tp sDfs sEx = do
             el "tr" $ do
                 elClass "th" "th-norm th-placing" $ text "✓"
                 elClass "th" "th-placing" $ text "Place"
-                elClass "th" "th-pilot" $ text "###-Pilot"
+                elClass "th" "th-pilot" . dynText $ ffor w hashIdHyphenPilot
                 elClass "th" "th-start-early" $ text "Early †"
                 elClass "th" "th-early-demerit" $ text "Points"
 
@@ -186,6 +187,7 @@ tableScorePenal hgOrPg early sgs _ln dnf' dfNt _vy vw _wg pt tp sDfs sEx = do
                 simpleList
                     sDfs
                     (pointRow
+                        w
                         (earliest <$> early)
                         (snd <$> cTimePoints)
                         (snd <$> cArrivalPoints)
@@ -194,7 +196,7 @@ tableScorePenal hgOrPg early sgs _ln dnf' dfNt _vy vw _wg pt tp sDfs sEx = do
                         tp
                         (Map.fromList <$> sEx))
 
-            dnfRows dnfPlacing dnf'
+            dnfRows w dnfPlacing dnf'
             return ()
 
         let tdFoot = elAttr "td" ("colspan" =: "15")
@@ -259,7 +261,8 @@ tableScorePenal hgOrPg early sgs _ln dnf' dfNt _vy vw _wg pt tp sDfs sEx = do
 
 pointRow
     :: MonadWidget t m
-    => Dynamic t JumpTheGunLimit
+    => Dynamic t Int
+    -> Dynamic t JumpTheGunLimit
     -> Dynamic t T.Text
     -> Dynamic t T.Text
     -> Dynamic t DfNoTrack
@@ -268,7 +271,7 @@ pointRow
     -> Dynamic t (Map.Map Pilot Norm.NormBreakdown)
     -> Dynamic t (Pilot, Breakdown)
     -> m ()
-pointRow earliest cTime cArrival dfNt pt tp sEx x = do
+pointRow w earliest cTime cArrival dfNt pt tp sEx x = do
     let pilot = fst <$> x
     let xB = snd <$> x
     let y = ffor3 pilot sEx x (\pilot' sEx' (_, Breakdown{total = p'}) ->
@@ -288,8 +291,8 @@ pointRow earliest cTime cArrival dfNt pt tp sEx x = do
     let jtg = jump . snd <$> x
     let jtgPenalties = penaltiesJump . snd <$> x
 
-    let classPilot = ffor2 pilot dfNt (\p (DfNoTrack ps) ->
-                        let n = showPilot p in
+    let classPilot = ffor3 w pilot dfNt (\w' p (DfNoTrack ps) ->
+                        let n = showPilot w' p in
                         if p `elem` (Pilot.pilot <$> ps)
                            then ("pilot-dfnt", n <> " ☞ ")
                            else ("", n))
@@ -332,10 +335,11 @@ pointRow earliest cTime cArrival dfNt pt tp sEx x = do
 
 dnfRows
     :: MonadWidget t m
-    => TaskPlacing
+    => Dynamic t Int
+    -> TaskPlacing
     -> Dynamic t Dnf
     -> m ()
-dnfRows place ps' = do
+dnfRows w place ps' = do
     let ps = unDnf <$> ps'
     len <- sample . current $ length <$> ps
     let p1 = take 1 <$> ps
@@ -345,20 +349,21 @@ dnfRows place ps' = do
         0 -> do
             return ()
         1 -> do
-            _ <- simpleList ps (dnfRow place (Just 1))
+            _ <- simpleList ps (dnfRow w place (Just 1))
             return ()
         n -> do
-            _ <- simpleList p1 (dnfRow place (Just n))
-            _ <- simpleList pN (dnfRow place Nothing)
+            _ <- simpleList p1 (dnfRow w place (Just n))
+            _ <- simpleList pN (dnfRow w place Nothing)
             return ()
 
 dnfRow
     :: MonadWidget t m
-    => TaskPlacing
+    => Dynamic t Int
+    -> TaskPlacing
     -> Maybe Int
     -> Dynamic t Pilot
     -> m ()
-dnfRow place rows pilot = do
+dnfRow w place rows pilot = do
     let dnfMajor =
             case rows of
                 Nothing -> return ()
@@ -388,7 +393,7 @@ dnfRow place rows pilot = do
     elClass "tr" "tr-dnf" $ do
         elClass "td" "td-norm td-placing" $ text ""
         elClass "td" "td-placing" . text $ showRank place
-        elClass "td" "td-pilot" . dynText $ showPilot <$> pilot
+        elClass "td" "td-pilot" . dynText $ ffor2 w pilot showPilot
         dnfMajor
         elClass "td" "td-total-points" $ text "0"
         dnfMinor