+++ lang-haskell/app-view/comp-view/FlareTiming/Task/Score/Split.hs
@@ -46,9 +46,9 @@ import WireTypes.Validity
     )
 import WireTypes.ValidityWorking (ValidityWorking(..), TimeValidityWorking(..))
 import WireTypes.Comp (UtcOffset(..), Discipline(..), MinimumDistance(..))
-import WireTypes.Pilot (Pilot(..), Dnf(..), DfNoTrack(..))
+import WireTypes.Pilot (Pilot(..), Dnf(..), DfNoTrack(..), pilotIdsWidth)
 import qualified WireTypes.Pilot as Pilot (DfNoTrackPilot(..))
-import FlareTiming.Pilot (showPilot)
+import FlareTiming.Pilot (showPilot, hashIdHyphenPilot)
 import FlareTiming.Task.Score.Show
 
 tableScoreSplit
@@ -69,6 +69,7 @@ tableScoreSplit
     -> Dynamic t [(Pilot, Norm.NormBreakdown)]
     -> m ()
 tableScoreSplit utcOffset hgOrPg free sgs _ln dnf' dfNt vy vw wg pt tp sDfs sEx = do
+    let w = ffor sDfs (pilotIdsWidth . fmap fst)
     let dnf = unDnf <$> dnf'
     lenDnf :: Int <- sample . current $ length <$> dnf
     lenDfs :: Int <- sample . current $ length <$> sDfs
@@ -298,7 +299,7 @@ tableScoreSplit utcOffset hgOrPg free sgs _ln dnf' dfNt vy vw wg pt tp sDfs sEx 
             el "tr" $ do
                 elClass "th" "th-norm th-placing" $ text "✓"
                 elClass "th" "th-placing" $ text "Place"
-                elClass "th" "th-pilot" $ text "###-Pilot"
+                elClass "th" "th-pilot" . dynText $ ffor w hashIdHyphenPilot
 
                 elClass "th" "th-distance-points" $ text ""
                 elClass "th" "th-norm th-distance-points" $ text "✓"
@@ -325,6 +326,7 @@ tableScoreSplit utcOffset hgOrPg free sgs _ln dnf' dfNt vy vw wg pt tp sDfs sEx 
                 simpleList
                     sDfs
                     (pointRow
+                        w
                         (snd <$> cTimePoints)
                         (snd <$> cArrivalPoints)
                         utcOffset
@@ -334,7 +336,7 @@ tableScoreSplit utcOffset hgOrPg free sgs _ln dnf' dfNt vy vw wg pt tp sDfs sEx 
                         tp
                         (Map.fromList <$> sEx))
 
-            dnfRows dnfPlacing dnf'
+            dnfRows w dnfPlacing dnf'
             return ()
 
         let tdFoot = elAttr "td" ("colspan" =: "21")
@@ -396,7 +398,8 @@ tableScoreSplit utcOffset hgOrPg free sgs _ln dnf' dfNt vy vw wg pt tp sDfs sEx 
 
 pointRow
     :: MonadWidget t m
-    => Dynamic t T.Text
+    => Dynamic t Int
+    -> Dynamic t T.Text
     -> Dynamic t T.Text
     -> Dynamic t UtcOffset
     -> Dynamic t MinimumDistance
@@ -406,7 +409,7 @@ pointRow
     -> Dynamic t (Map.Map Pilot Norm.NormBreakdown)
     -> Dynamic t (Pilot, Breakdown)
     -> m ()
-pointRow cTime cArrival _utcOffset _free dfNt pt tp sEx x = do
+pointRow w cTime cArrival _utcOffset _free dfNt pt tp sEx x = do
     let pilot = fst <$> x
     let xB = snd <$> x
 
@@ -441,8 +444,8 @@ pointRow cTime cArrival _utcOffset _free dfNt pt tp sEx x = do
 
     let points = breakdown . snd <$> x
 
-    let classPilot = ffor2 pilot dfNt (\p (DfNoTrack ps) ->
-                        let n = showPilot p in
+    let classPilot = ffor3 w pilot dfNt (\w' p (DfNoTrack ps) ->
+                        let n = showPilot w' p in
                         if p `elem` (Pilot.pilot <$> ps)
                            then ("pilot-dfnt", n <> " ☞ ")
                            else ("", n))
@@ -480,10 +483,11 @@ pointRow cTime cArrival _utcOffset _free dfNt pt tp sEx x = do
 
 dnfRows
     :: MonadWidget t m
-    => TaskPlacing
+    => Dynamic t Int
+    -> TaskPlacing
     -> Dynamic t Dnf
     -> m ()
-dnfRows place ps' = do
+dnfRows w place ps' = do
     let ps = unDnf <$> ps'
     len <- sample . current $ length <$> ps
     let p1 = take 1 <$> ps
@@ -493,20 +497,21 @@ dnfRows place ps' = do
         0 -> do
             return ()
         1 -> do
-            _ <- simpleList ps (dnfRow place (Just 1))
+            _ <- simpleList ps (dnfRow w place (Just 1))
             return ()
         n -> do
-            _ <- simpleList p1 (dnfRow place (Just n))
-            _ <- simpleList pN (dnfRow place Nothing)
+            _ <- simpleList p1 (dnfRow w place (Just n))
+            _ <- simpleList pN (dnfRow w place Nothing)
             return ()
 
 dnfRow
     :: MonadWidget t m
-    => TaskPlacing
+    => Dynamic t Int
+    -> TaskPlacing
     -> Maybe Int
     -> Dynamic t Pilot
     -> m ()
-dnfRow place rows pilot = do
+dnfRow w place rows pilot = do
     let dnfMajor =
             case rows of
                 Nothing -> return ()
@@ -536,7 +541,7 @@ dnfRow place rows pilot = do
     elClass "tr" "tr-dnf" $ do
         elClass "td" "td-norm td-placing" $ text ""
         elClass "td" "td-placing" . text $ showRank place
-        elClass "td" "td-pilot" . dynText $ showPilot <$> pilot
+        elClass "td" "td-pilot" . dynText $ ffor2 w pilot showPilot
         dnfMajor
         elClass "td" "td-total-points" $ text "0"
         dnfMinor